Extracting nonce, and fixing bean wiring problem for OAuth classes.
