+++ src/main/java/uk/gov/ch/developer/docs/controller/developer/UserCallbackController.java
@@ -8,9 +8,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import uk.gov.ch.developer.docs.oauth.IOauth;
-import uk.gov.ch.developer.docs.oauth.IdentityProvider;
 import uk.gov.ch.developer.docs.session.SessionService;
+import uk.gov.ch.oauth.IOauth;
+import uk.gov.ch.oauth.IdentityProvider;
 import uk.gov.companieshouse.logging.Logger;
 import uk.gov.companieshouse.logging.LoggerFactory;
 
@@ -30,14 +30,17 @@ public class UserCallbackController {
     @ResponseBody
     public String getCallback(@RequestParam("state") String state,
             @RequestParam("code") String code) {
-        sessionService.getSessionDataFromContext();
+        //LOGGER.info("Code:"+code);
+        //LOGGER.info("State:"+state);
+
+        //sessionService.getSessionDataFromContext();
 
         final Payload payload = oauth.oauth2DecodeState(state);
         final JSONObject jsonObject = payload.toJSONObject();
         final String returnedNonce = jsonObject.getAsString("nonce");
+        LOGGER.info("Nonce:" + returnedNonce);
         if (!oauth.oauth2VerifyNonce(returnedNonce)) {
             LOGGER.error("Invalid nonce value in state during oauth2 callback");
-
         }
         return "Found";
     }
