+++ src/main/java/uk/gov/ch/oauth/IOauth.java
@@ -1,4 +1,4 @@
-package uk.gov.ch.developer.docs.oauth;
+package uk.gov.ch.oauth;
 
 import com.nimbusds.jose.Payload;
