+++ src/main/java/uk/gov/ch/oauth/IdentityProvider.java
@@ -1,4 +1,4 @@
-package uk.gov.ch.developer.docs.oauth;
+package uk.gov.ch.oauth;
 
 import java.util.Base64;
 import java.util.Base64.Decoder;
@@ -12,10 +12,7 @@ import uk.gov.companieshouse.logging.LoggerFactory;
 public class IdentityProvider implements IIdentityProvider {
 
     private static final Logger LOGGER = LoggerFactory.getLogger("docs.developer.ch.gov.uk");
-    private static final String OAUTH_COMPANY_SCOPE_PREFIX =
-            "https://api.companieshouse.gov.uk/company/";
     private final static Decoder decoder = Base64.getDecoder();
-    //final private String base64Key;
     final private byte[] requestKey;
     final private String authorizationUri;
     final private String clientId;
@@ -24,7 +21,6 @@ public class IdentityProvider implements IIdentityProvider {
 
     @Autowired
     public IdentityProvider(final EnvironmentReader reader) {
-        // this.base64Key = reader.getMandatoryString("OAUTH2_REQUEST_KEY");
         requestKey = decoder.decode(reader.getMandatoryString("DEVELOPER_OAUTH2_REQUEST_KEY"));
         this.authorizationUri = reader.getMandatoryString("OAUTH2_AUTH_URI");
         this.clientId = reader.getMandatoryString("CHS_DEVELOPER_CLIENT_ID");
@@ -59,7 +55,7 @@ public class IdentityProvider implements IIdentityProvider {
     }
 
     public String getAuthorisationUrl(final String state) {
-        String sb = getAuthorizationUri()
+        return getAuthorizationUri()
                 + "?"
                 + "client_id="
                 + getClientId()
@@ -68,7 +64,6 @@ public class IdentityProvider implements IIdentityProvider {
                 + "&response_type=code"
                 + "&state="
                 + state;
-        return sb;
     }
 
     public String getAuthorizationUri() {
