+++ src/main/resources/application.properties
@@ -17,5 +17,6 @@ authentication.url=${home.url}/authentication
 manageApplications.path=dev-hub/manageApplications
 manageApplications.url=${home.url}/manage-applications
 callback.url=${home.url}/oauth2/user/callback
+signin.url=${home.url}/signin
 overview.path=dev-hub/overview
 overview.url=${home.url}/overview
